refactor main.js scene3d.js
