+++ src/js/main.js
@@ -1,74 +1,11 @@
 import 'bootstrap/dist/js/bootstrap.bundle.min.js'
 import paper from 'paper'
-import * as THREE from 'three'
-import { createGrid } from './floor2d.js'
-import { init3D, scene, camera, renderer } from './scene3d.js'
+import { createGrid } from './2d/floor2d.js'
+import { init3D, camera, renderer } from './3d/scene3d.js'
+import { convertPathsTo3D } from './3d/pathsTo3d.js'
 
 createGrid()
 init3D()
-
-let walls = []
-
-export { walls }
-
-function convertPathsTo3D() {
-  if (walls.length > 0) {
-    walls.forEach((wall) => scene.remove(wall))
-    walls = []
-  }
-
-  paper.project.activeLayer.children.forEach((item) => {
-    if (
-      item instanceof paper.Path &&
-      item.segments.length >= 2 &&
-      item.segments[0].handleIn.isZero() &&
-      item.segments[0].handleOut.isZero() &&
-      item.segments[1].handleIn.isZero() &&
-      item.segments[1].handleOut.isZero()
-    ) {
-      // Ensure item is a Path with at least 2 segments
-      // Check if it's a straight line (no handles, meaning no curve, so no verteces)
-      const startX =
-        (item.segments[0].point.x / paper.view.bounds.width) * 10 - 5
-      const startZ =
-        -(item.segments[0].point.y / paper.view.bounds.height) * 10 + 5
-      const endX = (item.segments[1].point.x / paper.view.bounds.width) * 10 - 5
-      const endZ =
-        -(item.segments[1].point.y / paper.view.bounds.height) * 10 + 5
-
-      const wallLength = Math.sqrt(
-        Math.pow(endX - startX, 2) + Math.pow(endZ - startZ, 2)
-      )
-      const wallHeight = 2.5 // Height of the walls
-      const wallThickness = 0.1 // Thickness of the walls
-
-      const wallGeometry = new THREE.BoxGeometry(
-        wallLength,
-        wallHeight,
-        wallThickness
-      )
-      wallGeometry.translate(0, wallHeight / 2, 0)
-      const wallMaterial = new THREE.MeshStandardMaterial({
-        color: 0x888888,
-        opacity: 0.5,
-        transparent: true,
-      })
-      const wall = new THREE.Mesh(wallGeometry, wallMaterial)
-
-      wall.position.x = startX + endX
-      wall.position.y = 0
-      // Paper.js: Y increases downward (typical web/screen coordinates) Three.js: Z increases upward (typical 3D coordinates)
-      wall.position.z = -(startZ + endZ)
-
-      const angle = Math.atan2(endZ - startZ, endX - startX)
-      wall.rotation.y = angle
-      wall.scale.set(2, 2, 2)
-      wall.name = 'wall'
-      scene.add(wall)
-      walls.push(wall)
-    }
-  })
-}
 
 // Switch between 2D and 3D modes
 const container2D = document.getElementById('container2D')
@@ -90,14 +27,9 @@ switchButton.addEventListener('click', () => {
 
 // Handle window resizing
 window.addEventListener('resize', () => {
+  console.log('resizing...')
   paper.view.viewSize = new paper.Size(window.innerWidth, window.innerHeight)
   camera.aspect = window.innerWidth / window.innerHeight
   camera.updateProjectionMatrix()
   renderer.setSize(window.innerWidth, window.innerHeight)
 })
-
-window.selectImage = (imageName) => {
-  const element = document.getElementById('dropdownMenuButton')
-  element.innerText = `${imageName} `
-  element.setAttribute('data-alt', `${imageName}`)
-}
